Adjust update check to check against currently served build
Turns out Ledger doesn't make builds from a new tag available immediately, and the website returns 403 (Forbidden) until it's made available. Now checking against the canonical URL published on their website instead. Will continue to check the repo and write a warning if a newer tag is available, to give a heads-up of an upcoming build.